antimalaria_backend/requirements.txt
cryptography==45.0.3
Django
django-allauth==65.9.0
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
filelock==3.18.0
gdown==5.2.0
gevent==25.5.1
greenlet==3.2.3
gunicorn==23.0.0
joblib==1.5.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
numpy==1.26.4
pandas==2.3.0
psycopg2-binary==2.9.10
PubChemPy==1.0.4
PyJWT==2.10.1
rdkit==2025.3.2
rdkit-pypi==2022.9.5
requests==2.32.3
scikit-learn==1.3.2
uvicorn==0.34.3
xgboost==2.1.4
